Add sample app showcasing shared element transitions
Note that the sample crashes when going back since
SeekableTransitionState from google's compose
and SeekableTransitionState from jetbrains compose
have a breaking API change between them, making it
impossible to call it at runtime.

This will be fixed as soon as Jetbrains compose
catches up with those breaking changes. This should
not be a concern of this library however in order to
support its users to use the provided
AnimatedContentScope